Completed update funtionality of edit form
Since the module management was not complete before this commit, we had
problems such as adding the items which were not in db after the user had
changed selection in modbox and also removing selection from db when the
user has removed it from the modbox. Also handeling the scenerio for the
program when the item is already and unchecked and is being deleted from
the db when it does not exit.

All these functionality and solution to these events are included in this
commit